requirements.txt
certifi==2022.9.24
click==8.1.3
colorama==0.4.6
dnspython==2.2.1
Flask==2.2.2
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
PyJWT==2.6.0
pymongo==4.3.2
Werkzeug==2.2.2
zipp==3.10.0
